feat(evm-core): add `IntoInstructionResult` trait and bound on `FoundryEvmFactory`
Move `IntoInstructionResult` from anvil to foundry-evm-core with impls
for `HaltReason`, `OpHaltReason`, and `TempoHaltReason`. Add
`HaltReason: IntoInstructionResult` bound to `FoundryEvmFactory` so
generic code can convert halt reasons to instruction results.
